backend/requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
pandas
beautifulsoup4
fastapi
uvicorn
chromadb
trafilatura